Fix pandas/numpy serialization (#21)
* Fix pandas/numpy serialization

A bug was introduced in 62a4d48d, which resulted in errors when
serializing multi-indices (as well as other complex index types). This
was caused by two underlying issues:

- False assumption that all indices can be reproduced exactly as
  ``pd.Index(index.values, name=index.name)``
- Partd failing to properly serialize object arrays containing only
  collections (e.g. an array of tuples).

This PR fixes these issues, along with a few other minor fixes:

- Serialize pandas indices separately from blocks
- Properly serialize numpy arrays containing lists/tuples
- Removes unneeded `decode` function for numpy serialization (msgpack
  can do this for us)
- Some pep8 fixes

* Update .travis.yml

- Drop 2.6, add 3.5
- Use conda-forge channel to get recent pandas/numpy for 3.3
- Use conda instead of pip for locket

* Remove python 3.3 as well

It seems python3.3 + pandas > 0.16 is no longer available on
conda/conda-forge.